+++ test/dummy/db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended to check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20120419183855) do
+ActiveRecord::Schema.define(:version => 20120717182827) do
 
   create_table "catalog_versions", :force => true do |t|
     t.integer  "original_record_id"