+++ test/dummy/test/unit/portlets/uses_helper_portlet_test.rb
@@ -1,0 +1,9 @@
+require File.join(File.dirname(__FILE__), '/../../test_helper')
+
+class UsesHelperTest < ActiveSupport::TestCase
+
+  test "Should be able to create new instance of a portlet" do
+    assert UsesHelperPortlet.create!(:name => "New Portlet")
+  end
+  
+end